refactor(logger): convert files to typescript
