--- src/__internal__/utils/logger/index.js
@@ -1,1 +1,0 @@
-export { default } from "./logger";
